refactoring file structure
