+++ scripts_python/general.py
